fix: 인터페이스 파일명 변경
